// src/Tooltip/types.ts
import { TippyProps } from '@tippyjs/react';

export interface TooltipProps extends TippyProps {
  disabled?: boolean;
  placement?:
    | 'top'
    | 'top-start'
    | 'top-end'
    | 'right'
    | 'right-start'
    | 'right-end'
    | 'bottom'
    | 'bottom-start'
    | 'bottom-end'
    | 'left'
    | 'left-start'
    | 'left-end';
}